# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FraudWatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteMe
      tags: [users]
      summary: Delete current user account
      responses:
        "204":
          description: Deleted

  # ─── Files ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        isBlocked:
          type: boolean
        totalUploads:
          type: integer
        totalAnalyses:
          type: integer
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, email, name, role, isBlocked, totalUploads, totalAnalyses, createdAt]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
...